move PrivateKeyGenerator to crypto module
