+++ jvm-libs/utils/build.gradle
@@ -18,7 +18,7 @@ plugins {
 }
 
 dependencies {
-  implementation(project(":crypto"))
+  implementation(testFixtures(project(":crypto")))
   implementation "io.libp2p:jvm-libp2p"
   implementation "tech.pegasys.teku.internal:p2p"
   implementation "build.linea.internal:kotlin"
